Reorganize bbox functions
